[ui/#9] 신곡 그리드 구현
